Sat, Mar 18, 2023, 4:10 PM -05:00

+++ problem-set/population.c
@@ -36,7 +36,7 @@ int main(void)
         int new_llamas = starting_size / 3;
         int dead_llamas = starting_size / 4;
 
-        total_llamas = total_llamas + new_llamas - dead_llamas;
+        total_llamas = total_llamas + (float) new_llamas - (float) dead_llamas;
     }
 
     // do
